feat: remove multiple accounts
